mcp-proton/requirements.txt
mcp[cli]
fastapi
uvicorn[standard]
python-jose[cryptography]
httpx
